Changed the files to use text2vec-transformers
