+++ requirements.txt
@@ -1,42 +1,1 @@
-appnope==0.1.3
-asttokens==2.4.0
-Authlib==1.2.1
-backcall==0.2.0
-certifi==2023.7.22
-cffi==1.16.0
-charset-normalizer==3.3.0
-comm==0.1.4
-cryptography==41.0.4
-debugpy==1.8.0
-decorator==5.1.1
-executing==2.0.0
-idna==3.4
-ipykernel==6.25.2
-ipython==8.16.1
-jedi==0.19.1
-jupyter_client==8.3.1
-jupyter_core==5.3.2
-matplotlib-inline==0.1.6
-nest-asyncio==1.5.8
-packaging==23.2
-parso==0.8.3
-pexpect==4.8.0
-pickleshare==0.7.5
-platformdirs==3.11.0
-prompt-toolkit==3.0.39
-psutil==5.9.5
-ptyprocess==0.7.0
-pure-eval==0.2.2
-pycparser==2.21
-Pygments==2.16.1
-python-dateutil==2.8.2
-pyzmq==25.1.1
-requests==2.31.0
-six==1.16.0
-stack-data==0.6.3
-tornado==6.3.3
-traitlets==5.10.1
-urllib3==2.0.6
-validators==0.22.0
-wcwidth==0.2.8
 weaviate-client==3.24.2
